Address feedback: rename function and move it to kubeUtils
